modify: rename files, change like on Login button path to main page
