+++ yoojin/src/components/App/App.jsx
@@ -2,7 +2,7 @@ import "./App.css";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 import Login from "../../pages/LoginPage";
 import Main from "../../pages/MainPage";
-import ChatPage from '../../pages/ChatPage';
+import ChatPage from "../../pages/ChatPage";
 import Topbar from "../Navbar/TopBar";
 import Profile from "../../pages/ProfilePage";
 
@@ -13,9 +13,10 @@ const App = () => {
       <BrowserRouter>
         <Switch>
           <Route exact={true} path="/" component={Login} />
-          <Route path="/main" component={Main} exact={true}/>
+          <Route path="/main" component={Main} exact={true} />
           <Route path="/main/profile" component={Profile} />
-          <Route path="/chat" component={ChatPage} />
+          <Route path="/chattings" component={ChatPage} />
+          <Route path="/chattings/chat" component={ChatPage} />
         </Switch>
       </BrowserRouter>
     </div>
